+++ packages/sdk/src/client.ts
@@ -8,7 +8,14 @@ import createAuthRefreshInterceptor, {
     AxiosAuthRefreshRequestConfig,
 } from "axios-auth-refresh";
 
-import { Auth, Log, DraftResource, Config, CloudQuery } from "./services";
+import {
+    Auth,
+    Log,
+    DraftResource,
+    Config,
+    CloudQuery,
+    Storage,
+} from "./services";
 
 class Client {
     private baseUrl: string;
@@ -137,6 +144,10 @@ class Client {
         return new CloudQuery(this);
     }
 
+    get storage(): Storage {
+        return new Storage(this);
+    }
+
     async call<D>(payload: {
         method:
             | "get"
@@ -152,13 +163,7 @@ class Client {
         skipAuthRefresh?: boolean;
         headers?: AxiosRequestHeaders;
     }): Promise<D> {
-        const { params, method, url, skipAuthRefresh, headers } = payload;
-        let { data } = payload;
-
-        data = {
-            ...data,
-            applicationClientId: this.clientId,
-        };
+        const { params, method, url, skipAuthRefresh, headers, data } = payload;
 
         const config: AxiosAuthRefreshRequestConfig = {
             baseURL: this.baseUrl,