+++ packages/cloud/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @pankod/refine-cloud
 
+## 6.0.0
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @pankod/refine-sdk@0.6.0
+
+## 5.0.0
+
+### Patch Changes
+
+-   Updated dependencies [[`e0dce4dda7`](https://github.com/pankod/refine/commit/e0dce4dda7ec2a2fa5a8619cb2747e68865b76ec)]:
+    -   @pankod/refine-sdk@0.5.0
+
 ## 4.0.2
 
 ### Patch Changes